New code to add audit information to object. Removed unnedded fields and refactored code
This release adds a created time, user and updated time and user to all
main data. It also defines three roles for users. It revamps some
screens and fixes hibernate issues

+++ src/main/java/com/damosais/sid/webapp/AttackersView.java
@@ -6,6 +6,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.tepi.filtertable.FilterTable;
 
 import com.damosais.sid.database.beans.Attacker;
+import com.damosais.sid.database.beans.User;
+import com.damosais.sid.database.beans.UserRole;
 import com.damosais.sid.database.services.AttackerService;
 import com.damosais.sid.webapp.windows.AttackerWindow;
 import com.vaadin.data.util.BeanItem;
@@ -58,7 +60,8 @@ public class AttackersView extends VerticalLayout implements View, ClickListener
     @Override
     public void buttonClick(ClickEvent event) {
         final Button button = event.getButton();
-        if (addEvent.equals(button)) {
+        final User user = ((WebApplication) getUI()).getUser();
+        if (addEvent.equals(button) && user.getRoles().contains(UserRole.EDIT_DATA)) {
             attackerWindow.setAddMode(this);
             getUI().addWindow(attackerWindow);
         } else {
@@ -67,7 +70,7 @@ public class AttackersView extends VerticalLayout implements View, ClickListener
             if (GraphicResources.EDIT_ICON.equals(button.getIcon())) {
                 attackerWindow.setEditMode(attackerToAlter, this);
                 getUI().addWindow(attackerWindow);
-            } else if (GraphicResources.DELETE_ICON.equals(button.getIcon())) {
+            } else if (GraphicResources.DELETE_ICON.equals(button.getIcon()) && user.getRoles().contains(UserRole.EDIT_DATA)) {
                 attackerService.delete(attackerToAlter);
                 refreshTableContent();
             }
@@ -141,10 +144,12 @@ public class AttackersView extends VerticalLayout implements View, ClickListener
         // Now we add the container
         container = new BeanItemContainer<>(Attacker.class);
         container.addNestedContainerProperty("country.name");
+        container.addNestedContainerProperty("createdBy.name");
+        container.addNestedContainerProperty("updatedBy.name");
         table.setContainerDataSource(container);
         // Now we define which columns are visible and what are going to be their names in the table header
-        table.setVisibleColumns(new Object[] { "name", "country.name", "type", EDIT_BUTTON, DELETE_BUTTON });
-        table.setColumnHeaders(new String[] { "Name", "Country", "Type", "Edit", "Delete" });
+        table.setVisibleColumns(new Object[] { "name", "country.name", "type", "created", "createdBy.name", "updated", "updatedBy.name", EDIT_BUTTON, DELETE_BUTTON });
+        table.setColumnHeaders(new String[] { "Name", "Country", "Type", "Created", "Created by", "Last update", "Last updated by", "Edit", "Delete" });
         table.setColumnAlignment(EDIT_BUTTON, CustomTable.Align.CENTER);
         table.setColumnAlignment(DELETE_BUTTON, CustomTable.Align.CENTER);
         // Now we refresh the content
